moved relevant files to reactComponents folder
