+++ src/reactComponents/Editor.js
