+++ src/reactComponents/Preview.js
