+++ src/reactComponents/readDefaultContent.js
